+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_cross_ios_test.dart
@@ -27,16 +27,21 @@ const name = 'add';
 ///
 /// | #   | Architecture | IOSSdk          | IOSVersion | Link Mode | Language    | Optimization Level |
 /// |-----|--------------|-----------------|------------|-----------|-------------|--------------------|
-/// | 1   | arm64        | iphoneos        | 16         | bundled   | c           | Os                 |
+/// | 1   | arm64        | iphoneos        | 16         | bundled   | objective c | Os                 |
 /// | 2   | arm64        | iphoneos        | 16         | static    | c           | unspecified        |
-/// | 3   | arm64        | iphoneos        | 17         | bundled   | c           | O3                 |
-/// | 4   | arm64        | iphonesimulator | 16         | static    | objective c | O0                 |
-/// | 5   | x64          | iphonesimulator | 17         | bundled   | objective c | O1                 |
-/// | 6   | x64          | iphonesimulator | 17         | static    | c           | O2                 |
+/// | 3   | arm64        | iphonesimulator | 16         | bundled   | objective c | O3                 |
+/// | 4   | arm64e       | iphoneos        | 17         | static    | c           | O2                 |
+/// | 5   | arm64e       | iphonesimulator | 17         | bundled   | objective c | O1                 |
+/// | 6   | x64          | iphonesimulator | 16         | static    | objective c | O0                 |
+/// | 7   | x64          | iphonesimulator | 17         | bundled   | c           | Os                 |
 final configurations =
     TestCaseSelector(
       dimensions: {
-        Architecture: [Architecture.arm64, Architecture.x64],
+        Architecture: [
+          Architecture.arm64,
+          Architecture.arm64e,
+          Architecture.x64,
+        ],
         IOSSdk: [IOSSdk.iPhoneOS, IOSSdk.iPhoneSimulator],
         IOSVersion: [
           IOSVersion.flutterHighestBestEffort,